Updated byeselinux to the one compiled by @McBane87, this is should fix the T2 Ultra Lollipop issues. Updated modulecrcpatch to the version statically compiled by @McBane87 Changed some code around in dualrecovery.sh making it easier to log more Updated sysrw.sh to only install the wp_mod.ko kernel module when needed (which is only if /sys/kernel/security/sony_ric is mounted in one of the init files) Updated dualrecovery.sh to unmount the mountpoint /system/odex instead of the device it should be mounted on, because that might negate the solution for the Z/ZL/ZR and TabZ. Updated TWRP to version 2.8.7.0

+++ includes/flashable/tmp/sysrw.sh
@@ -8,12 +8,16 @@
 
 BUSYBOX="/sbin/busybox"
 
-$BUSYBOX chmod 777 /tmp/modulecrcpatch
-for module in /system/lib/modules/*.ko; do
-	/tmp/modulecrcpatch $module /tmp/wp_mod.ko 1> /dev/null
-done
+if [ -e "/system/.XZDualRecovery/init.rc" -a "$($BUSYBOX grep '/sys/kernel/security/sony_ric/enable' /system/.XZDualRecovery/init.* | $BUSYBOX wc -l)" != "0" ]; then
 
-$BUSYBOX cp /tmp/wp_mod.ko /system/lib/modules/wp_mod.ko
-$BUSYBOX chmod 644 /system/lib/modules/wp_mod.ko
+	$BUSYBOX chmod 777 /tmp/modulecrcpatch
+	for module in /system/lib/modules/*.ko; do
+		/tmp/modulecrcpatch $module /tmp/wp_mod.ko 1> /dev/null
+	done
+
+	$BUSYBOX cp /tmp/wp_mod.ko /system/lib/modules/wp_mod.ko
+	$BUSYBOX chmod 644 /system/lib/modules/wp_mod.ko
+
+fi
 
 exit 0
